pyproject.toml
[tool.poetry]
name = "fastrepl"
version = "0.0.1"
description = "Simple Run-Eval-Polish Loop with LLM"
authors = ["yujonglee <yujonglee.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
litellm = "^0.1.392"
evaluate = "^0.4.0"
cuid2 = "^2.0.0"
graphviz = "^0.20.1"
rich = "^13.5.2"
iprogress = "^0.4"
ipywidgets = "^8.1.0"
sqlalchemy = "^2.0.19"
textual = { version = "^0.32.0", optional = true }

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-xdist = "^3.3.1"
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"
black = "^23.7.0"
torch = "^2.0.1"
python-dotenv = "^1.0.0"
html2text = "^2020.1.16"
nbconvert = "^7.7.3"
ipykernel = "^6.25.1"

[tool.poetry.group.example]
optional = true

[tool.poetry.group.example.dependencies]
llama-index = "^0.8.2.post1"
instructorembedding = "^1.0.1"
sentence-transformers = "^2.2.2"
langchain = "^0.0.265"

[tool.coverage.run]
omit = ["**/__init__.py"]

[tool.mypy]
exclude = [".venv"]
explicit_package_bases = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
markers = ["todo"]

[tool.poetry.extras]
tui = ["textual"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
